Organization
I moved around some of the code in index.css to common.css to ensure that the menu stays the same throughout every page the user clicks on. Also added the HTML files for Browse and AboutUs. I left out 'Top Audio' since it's still open to debate what we want to call that page.