AudioFile rebrand and menu optimization
Made some changes to the appearances on the login and registration menu's. There's still some refinement that needs to be done on the positioning of the buttons and such, but I'm pretty happy with where they are for now and will be moving on to completing the About Us section.